ТЗ16 исправления на проверку

+++ server/src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -28,21 +28,21 @@ public class BookingController {
     @PostMapping
     public Booking2Dto createBooking(@RequestBody @Valid BookingDto bookingDto, HttpServletRequest request) throws ValidationException {
         log.debug("Получен {} запрос {} тело запроса: {}", request.getMethod(), request.getRequestURI(), bookingDto);
-        Long userId = (long) request.getIntHeader("X-Sharer-User-Id");
+        long userId = request.getIntHeader("X-Sharer-User-Id");
         return bookingService.createBooking(bookingDto, userId);
     }
 
     @PatchMapping(value = "/{bookingId}")
     public Booking2Dto bookingApproved(@PathVariable Long bookingId, @RequestParam String approved, HttpServletRequest request) throws ValidationException {
         log.debug("Получен {} запрос {}", request.getMethod(), request.getRequestURI());
-        Long userId = (long) request.getIntHeader("X-Sharer-User-Id");
+        long userId = request.getIntHeader("X-Sharer-User-Id");
         return bookingService.updateApproving(bookingId, userId, approved);
     }
 
     @GetMapping(value = "/{bookingId}")
     public Booking2Dto getBookingById(@PathVariable Long bookingId, HttpServletRequest request) throws ValidationException {
         log.debug("Получен {} запрос {}", request.getMethod(), request.getRequestURI());
-        Long userId = (long) request.getIntHeader("X-Sharer-User-Id");
+        long userId = request.getIntHeader("X-Sharer-User-Id");
         return bookingService.getBookingById(bookingId, userId);
     }
 
@@ -52,7 +52,7 @@ public class BookingController {
                                               @Positive @RequestParam(name = "size", defaultValue = "20") Integer size,
                                               HttpServletRequest request) throws ThrowableException {
         log.debug("Получен {} запрос {}", request.getMethod(), request.getRequestURI());
-        Long userId = (long) request.getIntHeader("X-Sharer-User-Id");
+        long userId = request.getIntHeader("X-Sharer-User-Id");
         PageRequest pageRequest = PageRequest.of(from / size, size);
         return bookingService.getAllBookingsByUserId(state, userId, pageRequest);
     }
@@ -63,7 +63,7 @@ public class BookingController {
                                                    @Positive @RequestParam(name = "size", defaultValue = "20") Integer size,
                                                    HttpServletRequest request) throws ThrowableException {
         log.debug("Получен {} запрос {}", request.getMethod(), request.getRequestURI());
-        Long ownerId = (long) request.getIntHeader("X-Sharer-User-Id");
+        long ownerId = request.getIntHeader("X-Sharer-User-Id");
         return bookingService.getAllBookingsItemsByUserId(ownerId, state, PageRequest.of(from, size));
     }
 }
